requirements.txt
Flask
python-dotenv
azure-storage-blob 
azure-cosmos
azure-identity
